Refactor project structure and CMake configuration, remove unused UI components
- Restructured the project and updated the CMake file to conditionally activate FreeRTOS based on lv_conf.h settings.
- Removed UI files and C++ files related to SquareLine Studio to improve project stability.
- Added a global include.
- Updated the README.md to reflect the changes.
- Increased heap size for debugging with SDL.
