--- UI/ui.h
@@ -1,6 +1,0 @@
-#ifndef UI_H
-#define UI_H
-
-#define UI_ENABLE 1
-
-#endif
