+++ config/FreeRTOSConfig.h
@@ -21,7 +21,7 @@ extern uint32_t SystemCoreClock;
 #define configTICK_RATE_HZ                      ( ( TickType_t ) 1000 )
 #define configMAX_PRIORITIES                    ( 5 )
 #define configMINIMAL_STACK_SIZE                ( ( unsigned short ) 256 )
-#define configTOTAL_HEAP_SIZE                   ( ( size_t ) ( 256 * 1024 ) )  // 10 KB Heap
+#define configTOTAL_HEAP_SIZE                   ( ( size_t ) ( 512 * 1024 * 1024 ) )  // 512 MB Heap
 #define configMAX_TASK_NAME_LEN                 ( 10 )
 #define configUSE_TRACE_FACILITY                1
 #define configUSE_16_BIT_TICKS                  0
