+++ main/inc/glob.h
@@ -1,0 +1,17 @@
+#ifndef GLOB_H
+#define GLOB_H
+
+#include <stdlib.h>
+#include <stdio.h>
+#include <unistd.h>
+
+
+
+
+
+
+
+void freertos_main( void ); 
+
+
+#endif
